Attack Direction
the player's attack direction now corresspondesnsns with their movement direction.

rotation of the attack animation/sprite is in the works; up, right and down are behaving correctrly but the left rotation is still wonky. i was faced with a third blue screen from my pc and have now given up for the day. thank you.